requirements.txt
gunicorn
flask
flask-login
python-dotenv
opencv-python-headless
numpy
python-dateutil
